Create Public Room Solved
