+++ Chat rooms/config.py
@@ -1,0 +1,3 @@
+# config.py
+class Config:
+    SECRET_KEY = 'votre_clé_secrète'
