+++ Chat rooms/models/room.py
@@ -1,0 +1,4 @@
+rooms = [
+    {"name": "Room1", "type": "Privé", "description": "Salle privée pour les membres VIP", "created_by": "Admin"},
+    {"name": "Room2", "type": "Public", "description": "Salle publique pour tout le monde", "created_by": "Admin"}
+]
