+++ Chat rooms/static/styles.css
@@ -1,84 +1,41 @@
 body {
     font-family: Arial, sans-serif;
-    background-color: #f4f4f9;
-    color: #333;
-    margin: 0;
-    padding: 20px;
+    margin: 20px;
 }
 
 .alert {
     padding: 10px;
     margin-bottom: 20px;
-    border-radius: 5px;
-}
-
-.alert-error {
-    background-color: #ffebee;
-    color: #c62828;
-    border: 1px solid #c62828;
-}
-
-.alert-success {
-    background-color: #e8f5e9;
-    color: #2e7d32;
-    border: 1px solid #2e7d32;
-}
-
-h1, h2 {
-    color: #007bff;
-}
-
-form {
-    margin-bottom: 20px;
-}
-
-input, select, button {
-    padding: 8px;
-    margin: 5px;
-    border: 1px solid #ccc;
+    border: 1px solid transparent;
     border-radius: 4px;
 }
 
-button {
-    background-color: #007bff;
-    color: white;
-    cursor: pointer;
+.alert-success {
+    color: #3c763d;
+    background-color: #dff0d8;
+    border-color: #d6e9c6;
 }
 
-button:hover {
-    background-color: #0056b3;
+.alert-error {
+    color: #a94442;
+    background-color: #f2dede;
+    border-color: #ebccd1;
 }
 
 table {
     width: 100%;
     border-collapse: collapse;
-    margin-top: 20px;
 }
 
-table th, table td {
-    padding: 10px;
+table, th, td {
+    border: 1px solid #ddd;
+}
+
+th, td {
+    padding: 8px;
     text-align: left;
-    border-bottom: 1px solid #ddd;
 }
 
-table th {
-    background-color: #007bff;
-    color: white;
-}
-
-table tr:hover {
-    background-color: #f1f1f1;
-}
-
-a {
-    color: #007bff;
-    text-decoration: none;
-}
-
-a:hover {
-    text-decoration: underline;
-}
-
-#createRoomForm {
-    margin-top: 20px;
+th {
+    background-color: #f2f2f2;
 }
