+++ Chat rooms/templates/enter_code.html
@@ -2,29 +2,14 @@
 <html lang="fr">
 <head>
     <meta charset="UTF-8">
-    <title>Entrer le code - BLING</title>
+    <title>Rejoindre une salle privée - BLING Chat Rooms</title>
     <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
 </head>
 <body>
-    <h1>Entrer le code d'accès</h1>
-    
-    <!-- Afficher les messages flash -->
-    {% with messages = get_flashed_messages(with_categories=true) %}
-        {% if messages %}
-            {% for category, message in messages %}
-                <div class="alert alert-{{ category }}">
-                    {{ message }}
-                </div>
-            {% endfor %}
-        {% endif %}
-    {% endwith %}
-
-    <!-- Formulaire pour entrer le code -->
-    <form action="{{ url_for('join_room', room_name=room_name) }}" method="post">
-        <input type="text" name="room_code" placeholder="Entrez le code d'accès (4 chiffres)" required>
-        <button type="submit">Rejoindre la salle</button>
+    <h1>Rejoindre {{ room_name }}</h1>
+    <form action="{{ url_for('rooms.join_room', room_name=room_name) }}" method="post">
+        <input type="text" name="room_code" placeholder="Entrez le code d'accès" required>
+        <button type="submit">Rejoindre</button>
     </form>
-
-    <p><a href="{{ url_for('index') }}">Retour à la liste des salles</a></p>
 </body>
 </html>
