+++ Chat rooms/templates/index.html
@@ -8,7 +8,6 @@
 <body>
     <h1>BLING Chat Rooms</h1>
     
-    <!-- Afficher les messages flash -->
     {% with messages = get_flashed_messages(with_categories=true) %}
         {% if messages %}
             {% for category, message in messages %}
@@ -19,45 +18,40 @@
         {% endif %}
     {% endwith %}
 
-    <!-- Bouton de connexion/déconnexion -->
     {% if user %}
-        <p>Connecté en tant que <strong>{{ user }}</strong> | <a href="{{ url_for('logout') }}">Déconnexion</a></p>
+        <p>Connecté en tant que <strong>{{ user }}</strong> | <a href="{{ url_for('auth.logout') }}">Déconnexion</a></p>
     {% else %}
-        <p><a href="{{ url_for('login') }}">Se connecter</a></p>
+        <p><a href="{{ url_for('auth.login') }}">Se connecter</a></p>
     {% endif %}
 
-    <!-- Bouton pour afficher/masquer le formulaire de création -->
     {% if user %}
         <button id="toggleCreateRoomForm">Créer une nouvelle salle</button>
     {% else %}
-        <p>Vous devez <a href="{{ url_for('login') }}">vous connecter</a> pour créer une salle.</p>
+        <p>Vous devez <a href="{{ url_for('auth.login') }}">vous connecter</a> pour créer une salle.</p>
     {% endif %}
 
-    <!-- Formulaire pour créer une nouvelle salle (masqué par défaut) -->
     <div id="createRoomForm" style="display: none;">
         <h2>Créer une nouvelle salle :</h2>
-        <form action="{{ url_for('create_room') }}" method="post">
+        <form action="{{ url_for('rooms.create_room') }}" method="post">
             <input type="text" name="room_name" placeholder="Nom de la salle" required>
             <select name="room_type" id="roomType" required>
                 <option value="Public">Public</option>
                 <option value="Privé">Privé</option>
             </select>
             <div id="codeField" style="display: none;">
-                <input type="text" name="room_code" placeholder="Code d'accès (4 chiffres)" required>
+                <input type="text" name="room_code" placeholder="Code d'accès (4 chiffres)">
             </div>
             <input type="text" name="description" placeholder="Description" required>
             <button type="submit">Créer la salle</button>
         </form>
     </div>
 
-    <!-- Formulaire de recherche -->
     <h2>Rechercher une salle :</h2>
-    <form action="{{ url_for('search') }}" method="get">
+    <form action="{{ url_for('rooms.search') }}" method="get">
         <input type="text" name="query" placeholder="Rechercher une salle...">
         <button type="submit">Rechercher</button>
     </form>
 
-    <!-- Liste des salles disponibles -->
     <h2>Liste des salles :</h2>
     <table>
         <thead>
@@ -78,9 +72,9 @@
                 <td>{{ room['created_by'] }}</td>
                 <td>
                     {% if room['type'] == 'Privé' %}
-                        <a href="{{ url_for('join_room', room_name=room['name']) }}">Rejoindre (Code requis)</a>
+                        <a href="{{ url_for('rooms.join_room', room_name=room['name']) }}">Rejoindre (Code requis)</a>
                     {% else %}
-                        <a href="{{ url_for('join_room', room_name=room['name']) }}">Rejoindre</a>
+                        <a href="{{ url_for('rooms.join_room', room_name=room['name']) }}">Rejoindre</a>
                     {% endif %}
                 </td>
             </tr>
@@ -88,9 +82,7 @@
         </tbody>
     </table>
 
-    <!-- Script JavaScript pour gérer l'affichage du formulaire et du champ de code -->
     <script>
-        // Afficher/masquer le formulaire de création de salle
         document.getElementById('toggleCreateRoomForm').addEventListener('click', function() {
             const form = document.getElementById('createRoomForm');
             if (form.style.display === 'none') {
@@ -100,7 +92,6 @@
             }
         });
 
-        // Afficher/masquer le champ de code en fonction du type de salle
         document.getElementById('roomType').addEventListener('change', function() {
             const codeField = document.getElementById('codeField');
             if (this.value === 'Privé') {
